% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{genelevelStatistics}
\alias{genelevelStatistics}
\title{hits per gene
hits per expression level
editing proportions per gene}
\usage{
genelevelStatistics(posGR)
}
\description{
hits per gene
hits per expression level
editing proportions per gene
}
