% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{make_test}
\alias{make_test}
\title{countFiles <- countFiles #'remove rep 3 in ECT2 roots
print(basename(countFiles))
#' ----displaySampleTable----------------------------------------------------
sampleTable
#' ----makeecs, eval=TRUE----------------------------------------------------
#' ----sizeFactors1----------------------------------------------------------
#' ----estDisp1--------------------------------------------------------------
#' ----testForDEU1,cache=TRUE------------------------------------------------
#' ----estFC,cache=TRUE------------------------------------------------------}
\usage{
make_test(out_dir, design_vector = c(rep("control", 5), rep("treat", 5)),
  ncores = 1)
}
\description{
countFiles <- countFiles #'remove rep 3 in ECT2 roots
print(basename(countFiles))
#' ----displaySampleTable----------------------------------------------------
sampleTable
#' ----makeecs, eval=TRUE----------------------------------------------------
#' ----sizeFactors1----------------------------------------------------------
#' ----estDisp1--------------------------------------------------------------
#' ----testForDEU1,cache=TRUE------------------------------------------------
#' ----estFC,cache=TRUE------------------------------------------------------
}
